menu  all & update  and delete Table

+++ controllers/menuControllers.go
@@ -1,12 +1,13 @@
 package controllers
 
 import (
-	"fmt"
 	"net/http"
 	"time"
 
 	"github.com/ShahSau/culinary-bliss/database"
+	"github.com/ShahSau/culinary-bliss/helpers"
 	"github.com/ShahSau/culinary-bliss/models"
+	"github.com/ShahSau/culinary-bliss/types"
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
@@ -15,6 +16,14 @@ import (
 
 var menuCollection *mongo.Collection = database.GetCollection(database.DB, "menu")
 
+// @Summary Get all menus
+// @Description Get all menus
+// @Tags Global
+// @Accept json
+// @Produce json
+// @Success 200 {object} string
+// @Failure 500 {object} string
+// @Router /menu [get]
 func GetMenus(c *gin.Context) {
 	menus, err := menuCollection.Find(c.Request.Context(), bson.M{}, nil)
 
@@ -39,13 +48,17 @@ func GetMenus(c *gin.Context) {
 
 }
 
+// @Summary Get a menu
+// @Description Get a menu
+// @Tags Global
+// @Accept json
+// @Produce json
+// @Param id path string true "Menu ID"
+// @Success 200 {object} string
+// @Failure 500 {object} string
+// @Router /menu/{id} [get]
 func GetMenu(c *gin.Context) {
 	var menuID = c.Param("id")
-
-	if err := c.ShouldBindJSON(&menuID); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
 
 	var menu models.Menu
 
@@ -68,14 +81,29 @@ func GetMenu(c *gin.Context) {
 
 }
 
+// @Summary Create a menu
+// @Description Create a menu
+// @Tags Admin
+// @Accept json
+// @Produce json
+// @Param menu body types.Menu true "Menu object"
+// @Success 201 {object} string
+// @Failure 500 {object} string
+// @Router /menu [post]
 func CreateMenu(c *gin.Context) {
-	var reqMenu models.Menu
+	var incomingMenu types.Menu
 
-	if err := c.ShouldBindJSON(&reqMenu); err != nil {
+	if err := c.ShouldBindJSON(&incomingMenu); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
+	var reqMenu models.Menu
+
+	reqMenu.Name = incomingMenu.Name
+	reqMenu.Description = incomingMenu.Description
+	reqMenu.Start_Date, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
+	reqMenu.End_Date, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
 	reqMenu.ID = primitive.NewObjectID()
 	reqMenu.CreatedAt, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
 	reqMenu.UpdatedAt, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
@@ -83,27 +111,45 @@ func CreateMenu(c *gin.Context) {
 
 	defer c.Request.Body.Close()
 
-	menuCreated, err := menuCollection.InsertOne(c.Request.Context(), reqMenu)
+	_, err := menuCollection.InsertOne(c.Request.Context(), reqMenu)
 
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
 
-	c.JSON(http.StatusCreated, gin.H{"error": false, "message": "Menu created successfully", "data": menuCreated, "status": http.StatusCreated, "success": true})
+	c.JSON(http.StatusCreated, gin.H{"error": false, "message": "Menu created successfully", "data": reqMenu, "status": http.StatusCreated, "success": true})
 }
 
+// @Summary Update a menu
+// @Description Update a menu
+// @Tags Admin
+// @Accept json
+// @Produce json
+// @Param id path string true "Menu ID"
+// @Param menu body types.Menu true "Menu object"
+// @Success 200 {object} string
+// @Failure 500 {object} string
+// @Router /menu/{id} [put]
 func UpdateMenu(c *gin.Context) {
-	var reqMenu models.Menu
+	var reqMenu types.Menu
 
 	if err := c.ShouldBindJSON(&reqMenu); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
+	userEmail, _ := c.Get("first_name")
+	var isAdmin = helpers.IsAdmin(userEmail.(string))
+
+	if !isAdmin {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are not authorized to view this resource"})
+		return
+	}
+
 	menuId := c.Param("id")
 
-	menu := models.Menu{}
+	var menu models.Menu
 
 	err := menuCollection.FindOne(c.Request.Context(), bson.M{"menu_id": menuId}).Decode(&menu)
 
@@ -112,20 +158,11 @@ func UpdateMenu(c *gin.Context) {
 		return
 	}
 
-	if !reqMenu.End_Date.IsZero() || !reqMenu.Start_Date.IsZero() {
-		if !inTimeSpan(reqMenu.Start_Date, reqMenu.End_Date, time.Now()) {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "Menu start date and end date must be within the current date"})
-			return
-		}
-	}
-
-	menu.ID = reqMenu.ID
 	menu.Name = reqMenu.Name
 	menu.Description = reqMenu.Description
-	menu.Start_Date = reqMenu.Start_Date
-	menu.End_Date = reqMenu.End_Date
-	menu.Menu_id = menuId
 	menu.UpdatedAt, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
+	menu.Start_Date, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
+	menu.End_Date, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
 
 	_, err = menuCollection.UpdateOne(c.Request.Context(), bson.M{"menu_id": menuId}, bson.M{"$set": menu})
 
@@ -137,8 +174,25 @@ func UpdateMenu(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"error": false, "message": "Menu updated successfully", "data": menu, "status": http.StatusOK, "success": true})
 }
 
+// @Summary Delete a menu
+// @Description Delete a menu
+// @Tags Admin
+// @Accept json
+// @Produce json
+// @Param id path string true "Menu ID"
+// @Success 200 {object} string
+// @Failure 500 {object} string
+// @Router /menu/{id} [delete]
 func DeleteMenu(c *gin.Context) {
 	menuId := c.Param("id")
+
+	userEmail, _ := c.Get("first_name")
+	var isAdmin = helpers.IsAdmin(userEmail.(string))
+
+	if !isAdmin {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are not authorized to view this resource"})
+		return
+	}
 
 	_, err := menuCollection.DeleteOne(c.Request.Context(), bson.M{"menu_id": menuId})
 
@@ -148,12 +202,4 @@ func DeleteMenu(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{"error": false, "message": "Menu deleted successfully", "status": http.StatusOK, "success": true, "data": nil})
-}
-
-func MenuByRestaurant(c *gin.Context) {
-	fmt.Println("FoodByRestaurant")
-}
-
-func inTimeSpan(start, end, check time.Time) bool {
-	return start.After(check) && end.After(start)
 }
